# docs/swagger.yaml
basePath: /v1
definitions:
  main.APIError:
    properties:
      data: {}
      message:
        type: string
      requestid:
        type: string
      status:
        type: integer
    type: object
  main.Book:
    properties:
      author:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      price:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    required:
    - author
    - description
    - id
    - price
    - title
    type: object
  main.StatusResponse:
    properties:
      message:
        type: string
      requestid:
        type: string
      status:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    name: Jerome Amon
    url: https://learn.cloudmentor-scale.com/contact
  description: This provides a CRUD service on Book.
  license:
    name: MIT
    url: https://github.com/jeamon/demo-redis/blob/main/LICENSE
  title: Book Store API
  version: "1.0"
paths:
  /api/v1/books:
    post:
      description: Creates a book submitted and returns its ID.
      operationId: create-book
      parameters:
      - description: Book to create
        in: body
        name: Book
        required: true
        schema:
          $ref: '#/definitions/main.Book'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.APIError'
      security:
      - Bearer: []
      summary: Creates new book.
      tags:
      - Books
  /status:
    get:
      description: Get how long the application has been online.
      operationId: get-status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.StatusResponse'
      summary: Get the app status
      tags:
      - Books
swagger: "2.0"